Zeige durch klicken auf ein Artwork im artworks.html eine detaillierte Ansicht dieses Artworks an; Füge die Möglichkeit Bilder hochzuladen, sie in der Datenbank zu speichern und anschließend über den RestController zu ziehen, hinzu[VA-52]

+++ src/main/java/de/hdbw/webshop/service/UserService.java
@@ -50,8 +50,8 @@ public class UserService {
                 true);
     }
 
-    public RolesEntity getRolesEntity(int roleId) {
-        return rolesRepository.findById((long) roleId).orElseThrow(() -> new RoleNotFoundException(
+    public RolesEntity getRolesEntity(long roleId) {
+        return rolesRepository.findById(roleId).orElseThrow(() -> new RoleNotFoundException(
                 "RoleEntityDTO by Id " +
                         roleId +
                         " was not found"));
